History actions: the three that ask the user something leave common/
common/history_actions.c is batch machinery -- run an action over a list of images, with
undo and a report. Three of its functions were not that:

- dt_history_copy_parts() and dt_history_paste_parts_prepare() open the module-picker
  dialog and store the answer in the copy/paste proxy;
- delete_history_callback() is a GtkAccelGroup accelerator with a confirmation dialog.

They were the only reason the file included gui/hist_dialog.h, gui/gtk.h and
gui/actions/menu.h, and nothing in the batch machinery uses them. Moved to
gui/common/history_actions_gui.{c,h}, following the _gui suffix convention -- which is not
cosmetic here, an unsuffixed gui/common/history_actions.h would shadow the backend header
for every file that includes it by that name.

A relocation, not an inversion: both callers (gui/actions/edit.c, libs/history.c) already
live above this layer, so no handler slot is needed. The partial-paste apply/on_image/
on_list functions stay in the backend -- they read the selops the dialog filled, and never
open it themselves.

common/history_actions.c now has zero GTK tokens and no gui/ include, and its header no
longer drags <gtk/gtk.h> into everything that includes it.
